refactor pivot into its own dynamic load
