+++ explorer/templates/explorer/base.html
@@ -29,7 +29,7 @@
             <pre>
 npm run dev
             </pre>
-        <div>Then refresh this page, and you'll get all of your styles, JS, and hot-reloading!</div>
+        <div>Then refresh this page, and you'll get all of your styles, JS, and hot-reloading.</div>
         <div>If this is the first time you are running the project, then run:</div>
         <pre>
 nvm install
